code for parsing a dataset into BIDS components
first attempt at a script to be run from the command line that takes a path to a dataset as input and finds all the snirf files in that directory. Then for each snirf file checks the filename and all parent directories for information on the bids fields (new function in cedalion/io/bids). If it doesn't anything it uses None. Then adds this info as a row per snirf file in a csv file that is saved in the dataset directory.