make GLM basis GaussianKernels behave like in Homer
The meaning of t_pre in GaussianKernels differed from the implemenation
in Homer2. In cedalion the Gaussians were centered in [-t_pre, t_post]
with enough padding so that the Gaussian kernels could decay back to the
baseline. In Homer the first Gaussian is centered at -t_pre. In order
to avoid confusion for switching users the basis function class
GaussianKernels behaves now as in Homer. The implementation with
padding is available as GaussianKernelsWithTails.