use metric for z-ordering the lines in scalp_plot
This changes the order in which the lines in the scalp_plot are plotted.
It moves lines with small metric values to the bottom and lines with
larger metric values to the top. For example in signal quality plots
where a high metric value indicates good quality (e.g. snr), good
channels will be plotted on top of channels with worse quality.