Update motion_correct.py
we have an issue with a channel resulting in nan's when running through TDDR. That channel was saturated for most of its run and thus had a constant value, except for small segments that were not saturated. For whatever reason, this resulted in NANs in tddr. I made this one change as a quick fix, but it idn't work. MORE LATER.