faciliate time series plots from files processed by the BU pipeline (#108)
These are changes I asked Sung to make. Previously this would load a pkl file that had the recording container for one run. In the BU processing pipeline, we are now saving a pkl file with a list of list of recording containers structures as subjects and runs.  This modification can now handle both cases.

In the future this will evolve to instead use the snakemake framework Shannon is now developing where by results are saved after each processing rule in the derivatives folder. this function will be modifiedin the future to allow us to select which of the existing output files to view without having to relaunch the GUI